+++ SiteUtilProjects/OnlineVideos.Sites.doskabouter/Ard/ArdMediathekUtil.cs
@@ -3,19 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 
-using OnlineVideos.Helpers;
-
 
 namespace OnlineVideos.Sites.Ard
 {
-    public class ArdConstants
-    {
-
-        public static Uri API_URL { get; } = new Uri("https://api.ardmediathek.de");
-        public static string ITEM_URL { get; } = API_URL + "/page-gateway/pages/ard/item/";
-        public static int DAY_PAGE_SIZE { get; } = 100;
-    }
-
     public class ArdMediathekUtil : SiteUtilBase
     {
         public static readonly string PLACEHOLDER_IMAGE_WIDTH = "{width}";
