+++ SiteUtilProjects/OnlineVideos.Sites.doskabouter/OnlineVideos.Sites.doskabouter.csproj
@@ -67,10 +67,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Ard\ArdConstants.cs" />
     <Compile Include="Ard\ArdMediathekUtil.cs" />
     <Compile Include="Ard\ArdPartner.cs" />
-    <Compile Include="Ard\ArdWidgetTeaserDeserializerBase.cs" />
     <Compile Include="Ard\CollectionExtensions.cs" />
+    <Compile Include="Ard\ContinuationToken.cs" />
+    <Compile Include="Ard\Json\ArdCategoryDeserializer.cs" />
+    <Compile Include="Ard\Json\ArdMediaStreamsDeserializer.cs" />
+    <Compile Include="Ard\Json\ArdVideoInfoDeserializer.cs" />
+    <Compile Include="Ard\Json\ArdWidgetTeaserDeserializerBase.cs" />
     <Compile Include="Ard\Services.cs" />
     <Compile Include="Ard\Dto.cs" />
     <Compile Include="DailyMotionUtil.cs" />
